# Omit fields fix (for #82)