## Description 
Fixes #82
~~This version preserves the previous behavior when it would not have resulted in an error; this is when (number of omitted fields) == ((number of fields in struct) - (number of omitted fields)). So in that case the hashed tuple keys/field names are still the ones given to `omit_fields` rather than the ones left after omitting, but this way any hashes created with the previous version should be unchanged.~~

~~Up to you if that is preferable to just doing the "correct" thing everywhere which may require updating the HashVersion.~~

edit: Following the discussion below it was judged best to just fix the bug without bumping the HashVersion.

## Benchmarks
edit: Updating these after rebase, mostly because the new ones in the PR were run on a faster system.
### Before
```
16×6 DataFrame
 Row │ benchmark   hash       version    base        trait       ratio     
     │ SubStrin…   SubStrin…  SubStrin…  String      String      Float64   
─────┼─────────────────────────────────────────────────────────────────────
   1 │ dicts       crc        4          967.198 μs  55.187 ms   57.0584
   2 │ structs     crc        4          33.092 μs   545.551 μs  16.4859
   3 │ tuples      crc        4          32.972 μs   372.394 μs  11.2942
   4 │ numbers     crc        4          16.301 μs   145.906 μs   8.95074
   5 │ dataframes  crc        4          43.913 μs   322.088 μs   7.33468
   6 │ symbols     crc        4          705.683 μs  1.325 ms     1.87718
   7 │ missings    crc        4          213.153 μs  246.706 μs   1.15741
   8 │ strings     crc        4          901.003 μs  278.867 μs   0.309507
   9 │ dicts       sha256     4          1.814 ms    89.512 ms   49.353
  10 │ structs     sha256     4          773.171 μs  2.054 ms     2.65633
  11 │ tuples      sha256     4          772.570 μs  1.872 ms     2.42351
  12 │ symbols     sha256     4          2.209 ms    3.577 ms     1.61925
  13 │ numbers     sha256     4          386.430 μs  554.387 μs   1.43464
  14 │ dataframes  sha256     4          784.483 μs  1.081 ms     1.37797
  15 │ missings    sha256     4          645.860 μs  672.500 μs   1.04125
  16 │ strings     sha256     4          2.481 ms    2.425 ms     0.977336
```


### After
```
16×6 DataFrame
 Row │ benchmark   hash       version    base        trait       ratio     
     │ SubStrin…   SubStrin…  SubStrin…  String      String      Float64   
─────┼─────────────────────────────────────────────────────────────────────
   1 │ dicts       crc        4          948.822 μs  54.548 ms   57.4906
   2 │ structs     crc        4          32.171 μs   532.055 μs  16.5383
   3 │ tuples      crc        4          32.130 μs   364.188 μs  11.3348
   4 │ numbers     crc        4          16.140 μs   140.665 μs   8.7153
   5 │ dataframes  crc        4          42.881 μs   311.368 μs   7.26121
   6 │ symbols     crc        4          699.912 μs  1.325 ms     1.89366
   7 │ missings    crc        4          210.477 μs  238.461 μs   1.13296
   8 │ strings     crc        4          896.282 μs  260.171 μs   0.290278
   9 │ dicts       sha256     4          1.778 ms    88.896 ms   49.9836
  10 │ structs     sha256     4          772.479 μs  2.019 ms     2.61424
  11 │ tuples      sha256     4          772.539 μs  1.850 ms     2.39407
  12 │ symbols     sha256     4          2.184 ms    3.426 ms     1.56872
  13 │ dataframes  sha256     4          784.161 μs  1.068 ms     1.36218
  14 │ numbers     sha256     4          386.339 μs  513.720 μs   1.32971
  15 │ missings    sha256     4          636.372 μs  661.930 μs   1.04016
  16 │ strings     sha256     4          2.425 ms    2.392 ms     0.986562
```