@jrbp 2025-06-02
I pushed a commit updating the pluto compatibility to 0.20. So you might be able to just rerun CI.